requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.7
cloudinary==1.41.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
